chore(test): apply review fixes for monaco mock and vitest config
Extract the duplicated Monaco stub to a shared mock, omit maxWorkers
via conditional spread in CI, and drop a gitignored path from TODOS.
